Removed all the wrong "d eufoniche"
In Italian, the "d eufonica" is a "d" added to the conjunction "e" or to "a" when tehy appear right before words beginning with the same vowel. For example, rather than the wrong:

  professori e esperti
  provare a anticipare

one should write:

  professori ed esperti
  provare ad anticipare

The "d eufonica" cannot be put prior words beginning with different vowels. For example:

  ed inoltre

is wrong, and should be written, instead:

  e inoltre

This commit fixes all the wrong "d eufoniche", removing them when they are not justified. It also replaces "ad esempio" with "per esempio", for the same reason.